pyproject.toml
[project]
name = "hoopla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.57.0",
    "nltk==3.9.1",
    "numpy>=2.2.6",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]

[dependency-groups]
dev = [
    "ipython>=8.37.0",
    "pytest>=9.0.1",
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = [
    "tests"
]
pythonpath = [
    "."
]

[tool.basedpyright]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/.venv",
    "**/.[!/.]*/**"
]

[tool.pyrefly]
project-excludes = [
    "**/node_modules",
    "**/__pycache__",
    "**/*venv/**",
    "**/.[!/.]*/**"
]

[tool.ty.environment]
root = ["./cli"]
